requirements.txt
Django>=1.3.1
south>=0.7.3
django-annoying
django-sekizai
